+++ src/main/java/game/BallCounts.java
@@ -1,6 +1,7 @@
 package game;
 
 import constants.NumberConstants;
+import util.CountsMessage;
 
 public class BallCounts {
