+++ src/main/java/game/Option.java
@@ -1,4 +1,4 @@
-package client;
+package game;
 
 import java.util.Arrays;
