+++ src/main/java/game/Referee.java
@@ -1,6 +1,5 @@
 package game;
 
-import client.Ball;
 import constants.NumberConstants;
 import java.util.function.Function;
