+++ src/main/java/view/Output.java
@@ -1,4 +1,4 @@
-package util;
+package view;
 
 import constants.MessageConstants;
